% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte.R
\name{dayte}
\alias{dayte}
\title{Get Day of the year as a Date}
\usage{
dayte(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A Date or POSIXt object.
}
\description{
Returns the day of the year as a Date by setting the
year to be 1972 (as a leap year with two leap seconds)
}
\examples{
dayte(as.Date("2001-05-16"))
dayte(as.Date("2004-02-29"))
}
