% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte.R
\name{dayte_time}
\alias{dayte_time}
\title{Get Day and time of the year as a POSIX}
\usage{
dayte_time(x)
}
\arguments{
\item{x}{A POSIXt object.}
}
\value{
A Date object.
}
\description{
Returns the day of the year as a Date by setting the
year to be 1972 (as a leap year with two leap seconds)
}
\examples{
dayte_time(as.POSIXct("2001-05-16 02:03:04"))
}
