% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte.R
\name{doy2date}
\alias{doy2date}
\title{Day of the Year as an integer to a Date}
\usage{
doy2date(x, year)
}
\arguments{
\item{x}{The integer vector}

\item{year}{An integer vector of the year(s)}
}
\value{
A Date vector.
}
\description{
Day of the Year as an integer to a Date
}
\examples{
doy2date(1:2, 2000)
doy2date(1:2, 2002:2001)
}
