% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{ps_separate_date}
\alias{ps_separate_date}
\title{Separate Date}
\usage{
ps_separate_date(
  data,
  col = "Date",
  into = c("Year", "Month", "Day"),
  remove = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{col}{A column name or position.}

\item{into}{A character vector of length 3 specifying the names for the year, month and day components.}

\item{remove}{A flag specifying whether to remove the original column.}
}
\description{
Separates Date into Year, Month and Day.
}
\examples{
data <- data.frame(Date = Sys.Date())
ps_separate_date(data)
}
