% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{ps_separate_datetime}
\alias{ps_separate_datetime}
\title{Separate DateTime}
\usage{
ps_separate_datetime(data, col = "DateTime", into = c("Year", "Month",
  "Day", "Hour", "Minute", "Second"), remove = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{col}{A column name or position.}

\item{into}{A character vector of length 6 specifying the names for the year, month, day, hour, minute and second components.}

\item{remove}{A flag specifying whether to remove the original column.}
}
\description{
Separates DateTime into Year, Month and Day.
}
\examples{
data <- data.frame(DateTime = Sys.time())
ps_separate_datetime(data)
}
