% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc-offset.R
\name{ps_utc_offset}
\alias{ps_utc_offset}
\title{UTC Offset}
\usage{
ps_utc_offset(x)
}
\arguments{
\item{x}{A POSIXct object.}
}
\value{
The offsets as a numeric vector.
}
\description{
Calculates the offsets in hours between the time zones
of a POSIXct object and the coordinated Universal Time time zone.
}
\examples{
times <- c("2001-06-01 00:00:00", "2001-01-01 00:00:00")
times <- as.POSIXct(times, tz = "PST8PDT")
times
ps_utc_offset(times)
}
