% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{ps_get_season}
\alias{ps_get_season}
\title{Get Season}
\usage{
ps_get_season(x, seasons = c(Spring = 3L, Summer = 6L, Autumn = 9L, Winter =
  12L))
}
\arguments{
\item{x}{A Date or POSIXt vector}

\item{seasons}{A uniquely named integer vector of the first month of each season.}
}
\value{
An ordered factor of the seasons.
}
\description{
Returns an ordered factor the user specified seasons.
If the first month of the first season isn't January (1L), then
the last season is considered to wrap into the following year.
}
\examples{
ps_get_season(as.Date(c("2001-01-01", "2001-02-28", "2012-09-01", "2012-12-01")))
ps_get_season(as.Date(c("2001-01-01", "2001-02-28", "2012-09-01", "2012-12-01")),
season = c(Monsoon = 2L, `Dry Period` = 6L))
}
